Extract: `VoteDetails` into separate component
